project/requirements.txt
torch==1.12.1
torchaudio==0.12.1
librosa==0.9.2
numpy==1.24.0
pandas==1.5.3
scikit-learn==1.2.2
umap-learn==0.5.3
matplotlib==3.7.1
seaborn==0.12.2
sentence-transformers==2.2.2
jupyter==1.0.0
kagglehub[pandas-datasets]==0.1.1
kaggle==1.5.12
tqdm==4.65.0
soundfile==0.12.1
